move enum/types to reducers/map.ts
