+++ app/src/UI/Features/LegacyMap/LayerPicker/LpGlobalFilters/LpGlobalFilters.tsx
@@ -1,11 +1,11 @@
 import { FilterAlt, FilterAltOff } from '@mui/icons-material';
 import { IconButton } from '@mui/material';
-import { MapRecordsetLayerFilterCategory } from '../../helpers/functional/layer-definitions/reusable-layer-specifications';
 import TooltipWithIcon from 'UI/Reusable/TooltipWithIcon/TooltipWithIcon';
 import { useDispatch, useSelector } from 'utils/use_selector';
 import UserSettings from 'state/actions/userSettings/UserSettings';
 import './lpGlobalFilters.css';
 import { Fragment } from 'react';
+import { MapRecordsetLayerFilterCategory } from 'state/reducers/map';
 
 const LpGlobalFilters = () => {
   const MAP_FILTER_TOOLTIP_TEXT = 'Show or hide the record layers on the map. This setting applies to all record sets.';
